fix: remove taks, rescheduling, tasks' order

+++ lib/utils/task_manager.dart
@@ -91,7 +91,7 @@ class TaskManager {
     tasksDB.put(task.id, task);
     if (parentTask != null) {
       task.parentTask = parentTask;
-      parentTask.subtasks.add(task);
+      parentTask.subtaskIds.add(task.id);
       tasksDB.put(parentTask.id, parentTask);
     }
     logInfo('Created task: ${task.title}');
@@ -105,15 +105,21 @@ class TaskManager {
     return task;
   }
 
-  void deleteTask(Task task) {
+  void deleteTaskById(String taskId) {
+    final task = tasksDB.get(taskId);
+    if (task == null) {
+      logWarning('Task with id $taskId not found');
+      return;
+    }
+
     tasksDB.delete(task.id);
     final parentTask = task.parentTask;
     if (parentTask != null) {
-      parentTask.subtasks.remove(task);
+      parentTask.subtaskIds.remove(task.id);
       tasksDB.put(parentTask.id, parentTask);
     }
-    for (var subtask in List.from(task.subtasks)) {
-      deleteTask(subtask);
+    for (var subtaskId in task.subtaskIds) {
+      deleteTaskById(subtaskId);
     }
     for (var scheduledTask in List.from(task.scheduledTasks)) {
       for (var day in daysDB.values) {
@@ -178,20 +184,15 @@ class TaskManager {
   }
 
   void manageTasks() {
-    removeScheduledTasks();
+    removeScheduledTasks(); // TODO: check if this works correctly
     final tasks =
         tasksDB.values
-            .where((task) => task.frequency == null && task.subtasks.isEmpty)
+            .where((task) => task.frequency == null && task.subtaskIds.isEmpty)
             .where((task) => task.id != 'free_time_manager')
             .where(
               (task) => !task.category.name.toLowerCase().contains('event'),
             )
             .toList();
-
-    logInfo('Total numer of tasks ${tasks.length}');
-    for (var task in tasks) {
-      logInfo('${task.title} ${task.subtasks.length}');
-    }
 
     final justScheduledTasks = <ScheduledTask>[];
 
@@ -200,6 +201,7 @@ class TaskManager {
         tasks,
         justScheduledTasks,
       );
+
       if (taskUrgencyMap.isEmpty) {
         log('No tasks left to schedule');
         break;
@@ -441,38 +443,59 @@ class TaskManager {
   }
 
   bool _isOrderCorrect(Task task) {
-    if (task.order != null && task.order! > 0 && task.parentTask != null) {
-      return !task.parentTask!.subtasks.any(
-        (subtask) =>
-            subtask.order != null &&
-            subtask.order! < task.order! &&
-            subtask.scheduledTasks.isEmpty,
-      );
+    if (task.order == null || task.order! <= 1 || task.parentTask == null) {
+      return true;
     }
+
+    final parentTask = tasksDB.get(task.parentTaskId);
+    final parentSubtasks =
+        tasksDB.values
+            .where((t) => parentTask?.subtaskIds.contains(t.id) ?? false)
+            .toList();
+
+    for (var subtask in parentSubtasks) {
+      if (subtask.order! < task.order! && subtask.scheduledTasks.isEmpty) {
+        return false;
+      }
+    }
+
     return true;
   }
 
   void removeScheduledTasks() {
-    final now = DateTime.now();
     for (var day in daysDB.values) {
-      final dayDate = DateTime.parse(day.day);
-      if (dayDate.isBefore(now)) continue;
+      // final dayDate = DateTime.parse(day.day); //TODO: update this later
+      // if (dayDate.isBefore(now)) continue;
 
       final toRemove =
           day.scheduledTasks
               .where((st) => st.type == ScheduledTaskType.defaultType)
               .toList();
+
+      if (toRemove.isEmpty) continue;
+
+      logInfo(
+        'Removing ${toRemove.length} scheduled tasks from day ${day.day}',
+      );
       for (var scheduledTask in toRemove) {
-        day.scheduledTasks.remove(scheduledTask);
+        day.scheduledTasks.removeWhere(
+          (st) => st.scheduledTaskId == scheduledTask.scheduledTaskId,
+        );
+
         final task = tasksDB.get(scheduledTask.parentTaskId);
         if (task != null) {
-          task.scheduledTasks.remove(scheduledTask);
+          task.scheduledTasks.removeWhere(
+            (st) => st.scheduledTaskId == scheduledTask.scheduledTaskId,
+          );
+
           tasksDB.put(task.id, task);
+        } else {
+          logWarning('Task with id ${scheduledTask.scheduledTaskId} not found');
         }
       }
       daysDB.put(day.day, day);
     }
-    logInfo('Removed scheduled tasks after ${now.toIso8601String()}');
+    logInfo('Removed all scheduled tasks');
   }
 
   void removeScheduledTasksFor(Task task) {
@@ -506,7 +529,6 @@ class TaskManager {
   Future<List<Task>> breakdownAndScheduleTask(Task task) async {
     logInfo('Breaking down task: ${task.title} ${task.estimatedTime}');
 
-    // Изменяем тип с List<String> на List<Map<String, dynamic>>
     final subtaskDataList = await taskBreakdownAPI.breakdownTask(
       task.title,
       task.estimatedTime.toString(),
@@ -514,8 +536,6 @@ class TaskManager {
 
     if (subtaskDataList.isEmpty) {
       logWarning('No subtasks generated for task: ${task.title}');
-      logInfo('Scheduling parent task: ${task.title}');
-      scheduler.scheduleTask(task, userSettings.minSession, urgency: null);
       return [];
     }
 
@@ -525,7 +545,6 @@ class TaskManager {
     int order = 1;
 
     for (var subtaskData in subtaskDataList) {
-      // Извлекаем title и estimatedTime из Map
       final subtaskTitle = subtaskData['title'] as String;
       final estimatedTime = subtaskData['estimatedTime'] as int;
 
@@ -536,27 +555,18 @@ class TaskManager {
         estimatedTime: estimatedTime,
         deadline: task.deadline,
         category: task.category,
+        color: task.color,
         parentTask: task,
         order: order++,
       );
       tasksDB.put(subtask.id, subtask);
       subtasks.add(subtask);
-      task.subtasks.add(subtask);
+      task.subtaskIds.add(subtask.id);
     }
 
-    tasksDB.put(task.id, task);
-    scheduleSubtasks(subtasks);
+    manageTasks();
 
     return subtasks;
-  }
-// TODO: fix scheduling and order
-  void scheduleSubtasks(List<Task> subtasks) {
-    logInfo('Scheduling ${subtasks.length} subtasks');
-    subtasks.sort((a, b) => (a.order ?? 0).compareTo(b.order ?? 0));
-    for (var subtask in subtasks) {
-      scheduler.scheduleTask(subtask, userSettings.minSession, urgency: null);
-      logInfo('Scheduled subtask: ${subtask.title}');
-    }
   }
 
   /// Starts a task or subtask
@@ -620,7 +630,9 @@ class TaskManager {
     // Check if parent task should be completed too
     final parentTask = task.parentTask;
     if (parentTask != null) {
-      if (parentTask.subtasks.every((subtask) => subtask.isDone)) {
+      if (parentTask.subtaskIds.every(
+        (subtaskId) => tasksDB.get(subtaskId)!.isDone,
+      )) {
         completeTask(parentTask);
       }
     }
@@ -633,8 +645,11 @@ class TaskManager {
     List<TaskSession> allSessions = List.from(task.sessions);
 
     // Add sessions from subtasks
-    for (var subtask in task.subtasks) {
-      allSessions.addAll(getTaskSessions(subtask));
+    for (var subtaskId in task.subtaskIds) {
+      final subtask = tasksDB.get(subtaskId);
+      if (subtask != null) {
+        allSessions.addAll(getTaskSessions(subtask));
+      }
     }
 
     return allSessions;
@@ -645,8 +660,11 @@ class TaskManager {
     int total = task.getTotalDuration();
 
     // Add duration from subtasks
-    for (var subtask in task.subtasks) {
-      total += getTotalDuration(subtask);
+    for (var subtaskId in task.subtaskIds) {
+      final subtask = tasksDB.get(subtaskId);
+      if (subtask != null) {
+        total += getTotalDuration(subtask);
+      }
     }
 
     return total;
